fix: harden git-ops, rename shell-utils, remove dead code
- Add -- separator to setExecutable (consistent with clearExecutable)
- Validate branch names in fileExistsOnBranch (reject leading -)
- Rename shell-utils.ts → regex-utils.ts (only escapeRegExp remains)
- Replace dead gitOps fallback in GraphQLCommitStrategy with guard
